+++ BaaServiceKitTests/ProofTests/ProofModelTests.swift
@@ -3,45 +3,31 @@ import XCTest
 
 class ProofModelTests: XCTestCase {
     
-    override func setUp() {
-        super.setUp()
-    }
-    
-    override func tearDown() {
-        super.tearDown()
-    }
-    
     func testProofResponse() throws {
         let file = Bundle(for: type(of: self)).url(forResource: "PartialFakeProofResponse", withExtension: "json")!
         
-        do {
-            let data = try Data(contentsOf: file)
-            let proofResponse = try ChainpointProofResponse.jsonDecoder.decode([ChainpointProofResponse].self, from: data)
-            
-            // Proof reponse root object tests
-            XCTAssertEqual(proofResponse.count, 1)
-            XCTAssertEqual(proofResponse.first!.hashIdNode, "a4b52f60-7322-11e8-876e-0159403461ed")
-            XCTAssertEqual(proofResponse.first!.anchorsComplete!.count, 1)
-            XCTAssertEqual(proofResponse.first!.anchorsComplete!.first!, "cal")
-            
-            // Proof response proof tests
-            XCTAssertEqual(proofResponse.first!.proof!.context, URL(string: "https://w3id.org/chainpoint/v3"))
-            XCTAssertEqual(proofResponse.first!.proof!.type, "Chainpoint")
-            XCTAssertEqual(proofResponse.first!.proof!.hash, "1957db7fe23e4be1740ddeb941ddda7ae0a6b782e636a9e00b5aa82db1e84547")
-            XCTAssertEqual(proofResponse.first!.proof!.hashIdNode, "a4b52f60-7322-11e8-876e-0159403461ed")
-            XCTAssertEqual(proofResponse.first!.proof!.hashSubmittedNodeAt, "2018-06-18T18:08:46Z")
-            XCTAssertEqual(proofResponse.first!.proof!.hashIdCore, "a73c4f70-7322-11e8-b7ce-019b42e3a86a")
-            XCTAssertEqual(proofResponse.first!.proof!.hashSubmittedCoreAt, "2018-06-18T18:08:50Z")
-            
-            // Branches tests
-            XCTAssertEqual(proofResponse.first!.proof!.branches.count, 1)
-            XCTAssertEqual(proofResponse.first!.proof!.branches.first!.label, "cal_anchor_branch")
-            
-            print("ops: \(proofResponse.first!.proof!.branches.first!.ops)")
-        }
-        catch {
-            print("error: \(error)")
-        }
+        let data = try Data(contentsOf: file)
+        let proofResponse = try ChainpointProofResponse.jsonDecoder.decode([ChainpointProofResponse].self, from: data)
         
+        // Proof reponse root object tests
+        XCTAssertEqual(proofResponse.count, 1)
+        XCTAssertEqual(proofResponse[0].hashIdNode, "a4b52f60-7322-11e8-876e-0159403461ed")
+        XCTAssertEqual(proofResponse[0].anchorsComplete!.count, 1)
+        XCTAssertEqual(proofResponse[0].anchorsComplete![0], "cal")
+        
+        // Proof response proof tests
+        XCTAssertEqual(proofResponse[0].proof!.context, URL(string: "https://w3id.org/chainpoint/v3"))
+        XCTAssertEqual(proofResponse[0].proof!.type, "Chainpoint")
+        XCTAssertEqual(proofResponse[0].proof!.hash, "1957db7fe23e4be1740ddeb941ddda7ae0a6b782e636a9e00b5aa82db1e84547")
+        XCTAssertEqual(proofResponse[0].proof!.hashIdNode, "a4b52f60-7322-11e8-876e-0159403461ed")
+        XCTAssertEqual(proofResponse[0].proof!.hashSubmittedNodeAt, "2018-06-18T18:08:46Z")
+        XCTAssertEqual(proofResponse[0].proof!.hashIdCore, "a73c4f70-7322-11e8-b7ce-019b42e3a86a")
+        XCTAssertEqual(proofResponse[0].proof!.hashSubmittedCoreAt, "2018-06-18T18:08:50Z")
+        
+        // Branches tests
+        XCTAssertEqual(proofResponse[0].proof!.branches.count, 1)
+        XCTAssertEqual(proofResponse[0].proof!.branches[0].label, "cal_anchor_branch")
+        
+        print("ops: \(proofResponse[0].proof!.branches[0].ops)")
     }
 }
